reduce task retry count 3 -> 2; do not start adb when restarting
Rationale:

Failing tasks tend to hand due to
https://github.com/flutter/flutter/issues/5477. This is a quick fix to
unblock the dashboard when there are many failing tasks.

For some reason `adb start-server` sometimes fails to exit. This change
removes this command. It should be safe to only kill the server because
any `adb` command causes it to start automatically anyway.