Re-design RequestHandler API (#330)
On the Go backend, many API request endpoints handle
GET requests, not POST. The existing API of ApiRequestHandler
encoded that only POST requests were to be handled, so this
API needed an update.

The new API calls for subclasses to override `post()` and
`get()` directly, and those methods can now return a `Body`
object directly if they want to return a body in their
response. Properties such as the request, response, auth
info, etc. are now held in Zone values and accessed through
the `RequestHandler` API via getters.

This also splits the authentication handling into a dedicated
class to make the code more testable.