Remove signing helloworld app check. (#680)
When devicelab was having signing issues we added a health check that
created and app, build it and sign it. Later on we had some issues
where a problem with flutter sdk cause all the test beds to get
quarantined as soon as it was running test.

App signing has been very stable recently and we don't need this health
check anymore. We are removing it to get a couple of minutes of
execution time back.

Bug:
https://github.com/flutter/flutter/issues/52286