Set github status for builds to execute immediately.
Currently the github build state is set after we get the first pubsub
notification from luci. Under heavy use condition this can be anytime
from a few seconds to over an hour.

This is misleading as a PR with empty state shows as green and
developers and bots may submit the changes even though they haven't been
tested yet.

Bugs:
  https://github.com/flutter/flutter/issues/53121
  https://github.com/flutter/flutter/issues/43524