Only process check updates if messages come from a supported repo.
The implementation of the checks api will be one repo at a time and we
need to support repos with checks api support and the ones no yet
migrated. This PR checks is the status message comes with all the
expected fields to decide whether to process checks api updates or not.

Bug:
    https://github.com/flutter/flutter/issues/56422