Move the github client auth closer to where it will be used.
Github client is now using short lived tokens and in some cases when we
need to validate is builds exist and cancel them the token expired in
the middle of the request. We are moving the github client creation and
authentication to the place it will be used instead of creating at the
beginning of the method.

Bug:
  https://github.com/flutter/flutter/issues/62233