Show which tasks are failing in the status bar (#1052)
When I was gardener last week, I was frustrated several times because Cocoon would say that a task was failing, but all of the tiles on the screen were either green or yellow, so I had to wait a while until something turned red to find out what was broken.

This adds information to the status display that shows which tasks are failing when the build is failing, so that even before any indication is given about the failure, you can tell that it failed.

The main changes in this PR are to add failed task information to the server response for build-status, and to display those failures in the UI in build_dashboard_page.dart. To do this, I created a BuildResponseProto so that it could return more than just a bool.

One alternative to this approach (or an additional thing that could be done) would be to introduce another color for "failed once, but still in progress" tasks (orange?) that would show up right away.