Add command line options for switching data source. (#1056)
I added a couple of simple command line options to allow you to run a debug version with production data, and a release version with a local server. They're --use-production-service and --no-use-production-service. They default to production on release and fake on debug, as before.

In order for the command line options to be passed to the Dart main, I updated the platform shims to the latest versions (I didn't touch web, because I was too afraid that would break something), which accounts for the bulk of this change. You can run the local app this way with flutter run -a --use-production-service, for instance. Sadly, for now this only works on non-web platforms.

The only substantive changes in this PR are in app_flutter/lib