Extract and test framework auto-labelling (#1141)
Applies a similar refactoring to the framework auto-labelling for files
as was done in #1140 for the engine. Also adds more test coverage for
framework labelling.

This also fixes a bug where PRs that included files with the word
'accessibility' in their path were missed due to a typo in the path
matching.