Updates status calculation to pass if any success.
Retry logic in flutter dashboard does not allow to retry if there was at
least a successful execution and the logic to paint the task-commit in
the grid was using the signal of the last task. This PR updates the
logic to calculate the grid status to make it consistent with the retry
logic.

Bug: https://github.com/flutter/flutter/issues/88197