Fix message when retargetting to default branch
The message posted when a PR is automatically retargetted to the default
branch used a variable for the original target, but incorrectly still
hard-coded "master" as the default brach name. This means that the
message is wrong, and is particularly confusing for the (currently
common, as repos are being transitioned) case where the bot is switching
a PR from 'master' to 'main', as what it actually says is that the PR
wasn't opened against 'master' so is being switched to 'master', but if
that's not correct it can be switched back to 'master' again.

This fixes the message to get the actual default branch for the
repository, rather than assuming 'master'.