Adding a filter for staging tasks
Adding a checkbox to menu for staging tasks (tasks whose names
are prefixed with 'Staging_build_').
By default staging tasks are not displayed.
Should have no other visible changes/functionality of the dashboard.

Note that due to bug #103082 this change is broken, because the
bug causes the checkboxes to not work.

Bug:103081