Exempt analysis_options.yaml from test requirement
Analysis options are inherently tested by steps that run analysis, so
don't need tests.

Extracts the common set of test-exempt files from the flutter/flutter
and flutter/plugins+packages checks to minimize duplicate code.
(Switching flutter/engine to sharing the same set of exemptions is out
of scope for this refactor, as it would change behavior.)