Teach test bot about Pigeon tests (#2333)
* Teach test bot about Pigeon tests

The `pigeon` package in flutter/packages has a very unusual test
structure, with several plugins and a common shared library that are all
just tests for generated Pigeon code. This causes a lot of false
positives with the test bot, so this adds the `platform_tests` directory
to the list of known tests.

* Autoformat