Use locking on cache calls that come through the CacheRequestHandler Endpoint in Cocoon (#2616)
We ocassionally observe a 500 error when accessing cache values as the get performs both a get and set action. If the value is not present in the cache then a delegate will collect the necessary data and then the value will be populated into the cache. This I suspect is leading to a condition with Redis where we get a "Bad state: StreamSink is bound to a stream" error. 

This code relies upon neat-cache which already has a bug open but I believe that this could protect cocoon from seeing the issue.

*List which issues are fixed by this PR. You must list at least one issue.*

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*