[codesign] look for cert in designated keychain (#2627)
Fix (hopefully) an inexplicable bug when codesign tool is used on chromium bots. Specify 'build.keychain' as the keychain to look for flutter identity.

Context: on luci runs, the flutter identity is correctly imported under 'build.keychain', and is listed as both 'matching identity' and 'valid identity' for codesign, and 'build.keychain' is set and verified to be the default keychain. However, when signing on chromium bots, there is this inexplicable flake (~30 percent of time) where the bots keep looking for flutter cert under 'login.keychain'. 

Open to suggestion:  if this works, we could supply '--keychain' as a parameter instead of hard coding it to be 'build.keychain'. And we can supply 'build.keychain' as a parameter on the recipe side.

I have tested the correctness of '--keychain' parameter locally: 1. remove flutter cert from login.keychain 2. add flutter cert to build.keychain 3. set build.keychain as default keychain. 4. codesign with '--keychain=build.keychain' argument would succeed.