Follow up Pull Request Revert issues are no longer created (#2637)
This pull request removes the feature of creating a review issue when a revert pull request is generated. Previously a tracking issue was created so that a flutter-hacker member would follow up on a review of the reverted request. This has not been a popular feature and has caused more of a headache than anything else.

*List which issues are fixed by this PR. You must list at least one issue.*
Feedback the devs.

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*