Missing async keyword not allowing cocoon to wait for build collection (#2642)
A change was made to collect the try tasks and cancel them for a pull request when it is updated, closed, or merged. The function that gets the tasks was missing the 'async' keyword which caused us not to wait for the results of the task collection. It would only work rarely but it should be collecting all of the targets for the pull request each time unless nothing was truly scheduled or run yet.

*List which issues are fixed by this PR. You must list at least one issue.*
Fixes [125278](https://github.com/flutter/flutter/issues/125278)

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*