[cocoon] Remove user authentication when backend calls fail. (#2646)
This PR updates the Build Dashboard to make sure its UI reflects the authentication state of the user more directly. The UI refreshes as "logged out" when calls to the backend that require a valid authentication fail.

Before this, it was not entirely clear why some actions failed (failing operations only raised a small snackbar at the bottom of the page, and there was no way to renew the user session other than full-page reloading). The app could be in an inconsistent state where it *seemed* to be logged-in, but the user authorization would have expired.

After this PR, the app will seemingly "kick the user out" when they attempt something that they don't have permission to do, so they can interactively re-authenticate. It will also disable the problematic buttons so they cannot be spammed once the backend deauthorizes a user.

The changes are:

* Stops asking "Google Sign In" plugin if the current user `isSignedIn`, it has no idea.
* Makes the `isAuthenticated` check local to Cocoon, and not async (`return user != null`)
  * Updates mocks (and tests) to account for the API changes.
* Adds a `clearUser` method to refresh the UI but not tell the plugin that the user has signed out, so re-login is faster.
  * Calls `clearUser` whenever any request to the server that has `idToken` fails.
    * Assumes that requests only fail for authentication reasons, which might not always be the case! Opportunity for improvement here!
    * Currently this applies only to two requests: `refreshGitHubCommits` and `rerunTask`.
* Minimizes the chances of users sending unauthorized requests to the server.
  * Disables the "RERUN" and "Refresh GitHub Commits" buttons when the user is deemed not authenticated, so they must re-login rather than being able to click forever.
  * Removes the logic that did `signIn` if `idToken` was missing. It doesn't work as intended anymore. Instead, harden the `idToken` getter so it *requires* `isAuthenticated`.
* Adds tests for all of the above.

## Issues

* Mitigates/Fixes https://github.com/flutter/flutter/issues/125171