Fix review approval check ordering (#2648)
It was found that a review author can request changes be made then later approve the pull request. However Autosubmit was not honoring the chronological order of the reviews and was incorrectly blocking reviews that succeeded the original changes requested review. We now are traversing the reviews in reverse chronological order.

Also added a check so the pr author cannot review his own pull request.

*List which issues are fixed by this PR. You must list at least one issue.*
Fixes https://github.com/flutter/flutter/issues/125280

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*