[dashboard] Remove View Logs button (#2812)
This button is no longer used in LUCI world

A few new users to the build dashboard were confused by the view logs button. They expected it would show a link of all the builds for that commit, but it redirects to the builder page. This can already be done via the column header icon.