Remove fake viewport from dashboard (#2838)
This removes the _FakeViewport from the build dashboard implementation. It was added in https://github.com/flutter/cocoon/pull/741, and we think it's sole purpose was the preserve the 1D expectations of widgets in the framework like scrollbars and overscroll indicators. Since adding support for 2D in the framework, the framework has been updated so that the fake viewport should no longer be necessary.

All of the tests pass, so I think this is safe to remove now that the framework is 2D compatible.

This came up as a test failure in https://github.com/flutter/flutter/pull/128812, where the function signature of getOffsetToReveal is being changed.