Do not dispose of the client in test check (#2854)
When requests are made to the webhook_subscription there is one method that when called is disposing of the github client. In the _checkForTests methods it calls dispose on the github client which closes the connection while it is still in scope. This is the only part of the code that does this. This was killing the client for any pull request event for edited, opened, and reopened but not synchronized. 

The fix will be investigated further but this prevents .ci.yaml checks and other issues blocking checkRuns from starting.

*List which issues are fixed by this PR. You must list at least one issue.*
Fixes: https://github.com/flutter/flutter/issues/129134

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*