Refactor the Validation Service classes so they only process one type of PR. (#2925)
This is another part of the Revert implementation that will simplify the implementation and (later) separation while reducing interruption when revert is deployed as it will be deployed to it's own endpoint. 

While this does look like a lot of file changes, it is mostly existing logic refactored to accomadate the separate paths through the code.

This change simply separates the regular pull request validation path and the revert validation path. It keeps the original logic as the updated  Revert logic in a different branch that will rebased on this change. As it is currently Revert will need to be able to handle multiple passes through the Autosubmit due to the way Github generates events when labels are added, etc...

*List which issues are fixed by this PR. You must list at least one issue.*
Part of https://github.com/flutter/flutter/issues/113867

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*