[dashboard + branch] reimplement branch fetching logic (#2924)
For branches on cocoon dashboard, fetch any branch that has recent activities, AND, all the release branches. 

Sorry for the delay, it took me a huge amount of time to restudy the branching logic and dart syntax. The mocks and stubs in tests took me quite a while to get right.