Insert a new commit on branch creation (#3005)
Issue: https://github.com/flutter/flutter/issues/132066

This is based on comments from https://github.com/flutter/cocoon/pull/2991, but the code was significantly different so it made more sense to create a new branch.

When a branch is created, get the associated commit and push it to the datastore with the new branch. This will allow dart-internal builders, which run on a new candidate branch creation, to have a commit to save their task results to.