Treat files of type `*test.py` as test files in the engine repo. (#3000)
I should also note that when I originally wrote the test, my test reported a false positive. This is because the `generateGithubWebhookMessage` utility was using the `master` branch as the default base branch for all repositories, even though some are actually `main` and some are `master`. This fixes that as well, so that the default value for the base branch in the generated webhook messages matches the actual default base branch for the specified repository.