Automatic Revert (Part 1): Base services needed for the automated revert. (#3083)
This is the first part of the changes for Automated revert. These are the base classes that actually perform the revert. 

There are two methods that can be used. Unfortunately the github graphql method does not currently work. I have an open issue with github but included the code here anyway.

The git command line tools currently perform the revert. It checks out the repo, reverts the commit on the command line and then pushes the commit to github. It then creates a branch with the change which will be auto approved by the fluttergithubbot. Then ultimately merged.

*List which issues are fixed by this PR. You must list at least one issue.*
Part of https://github.com/flutter/flutter/issues/113867

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*