Automatic Revert (Part 2): Add config support for automated revert (#3084)
The files contained here are support files needed to initially support the automated revert. There will be another pull request to add some of the configs to the global configuration file but this is safer to do.

We need to setup autosubmit so that it is automatically approved when it creates a revert request. 

*List which issues are fixed by this PR. You must list at least one issue.*
Part of https://github.com/flutter/flutter/issues/113867

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*