Automated Revert (Part 4): The automatic revert functionality (#3089)
This is the code that performs the revert automatically.

There are two paths in this code which is mainly due to the events generated by github. As a result of this there are two supported labels, 'revert' and 'revert of.'

Labels:
'revert' tells the service that this is a closed and merged change that will need to be reverted.
'revert of' tells the service this is a new pull request that reverts an already merged change. This pull request is open and unmerged.

In order to revert a change, the following will happen:
1st pass
1. Dev adds the revert label to a closed and merged pull request.
2. The webhook validates the request and sends it to pubsub.
3. The revert service will get the message and if the pull request is not older than 24 hours then it can be reverted.
4. The code will perform a sparse checkout (this does not check out the full source code only the files in the root).
5. A branch will be created with the commit hash
6. The change will be reverted and pushed to Github
7. 'revert' label is removed (This is something that currently has to happen due to the amount of events github generates which can cause duplicate processing)

2nd pass
1. Then the pull request will be opened and the 'revert of' label will be added.
2. This will trigger the second path in the code.
3. Validations will be run against the change according to the configuration for reverts.
4. fluttergithubbot will approve the pull request.
5. pull request will be merged.

Sample PR that is a revert generated by this service: https://github.com/flutter/cocoon/pull/3066
![image](https://github.com/flutter/cocoon/assets/32242716/09f8df7c-4964-42ca-b368-01091224fcda)

Note that this PR depends on the following pull requests:
https://github.com/flutter/cocoon/pull/3083
https://github.com/flutter/cocoon/pull/3084
https://github.com/flutter/cocoon/pull/3085

*List which issues are fixed by this PR. You must list at least one issue.*
Part of https://github.com/flutter/flutter/issues/113867

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*