Allow body and title to be null from original pull request when creating a revert request (#3107)
Occasionally devs may forget to add a body on a pull request, which caused a null operator on null value error preventing a revert from being created automatically. This change allows the title and body of the original pull request to be null. We then substitute that no value was provided in the revert request.

*List which issues are fixed by this PR. You must list at least one issue.*
Fixes https://github.com/flutter/flutter/issues/135536

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*