Revert not correctly collecting default branch (#3110)
This is a fix to the way the default branch is collected. The code was wrongly checking for if the default branch was empty after the config was collected. It was never empty because a default value of main was assigned so that was used each time.

This changes the default branch to have a default value of 'default' and then we check against that value to get true default branch from the repo.

Example pull request of the fix: https://github.com/flutter/flutter/pull/135692

*List which issues are fixed by this PR. You must list at least one issue.*
Fixes https://github.com/flutter/flutter/issues/135670

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*