Do not schedule targets on a revert (#3122)
Since we do not care about waiting for targets to complete on revert requests we can ignore them and save some capacity on these requests. 

We only require the ci.yaml validation check run on reverts. Not that if we change requirements this will need to be reworked.

Another option is to mark targets as run on revert or something to that effect and let autosubmit choose which to enforce.

*List which issues are fixed by this PR. You must list at least one issue.*
Fixes https://github.com/flutter/flutter/issues/135790

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*