Default engine to requiring tests (#3248)
The current engine rules are the reverse of the framework and packages repos; those assume that any file that has not been explicitly exempted needs tests, whereas the engine only requires tests for a specific set of files. This is prone to false negatives (e.g., if/when we start adding Swift files, someone would have had to remember to come update this rule, which probably wouldn't have happened).

This flips the engine to be consistent with the other repositories. This will almost certainly cause a lot of false positives in the short term while the set of files that should be exempted is worked out.

Also fixes a bug where the removal-only case was only being applied to Dart files, not other languages.