Avoid completing a non-nullable Completer with a nullable value (#3328)
Implicitly passing a nullable value to a Completer of a non-nullable type can lead to surprising null-asserting exceptions. See https://github.com/dart-lang/sdk/issues/53253 for more details. In this PR, we add an explicit null-assertion, which makes this call in-line with the general concepts of null safety.

Fixes https://github.com/flutter/flutter/issues/137294