Add tags to revert message for later processing (#3523)
Sending information to discord will be much easier with this feature. We will send this message to discord after the 'revertOf' request has been merged, which means we will need to collect some information from the pull request description. The tags created will not be rendered in the markdown but will be present in the string body.

This allows us to not have to use the cache and not have to rely on state information.

*List which issues are fixed by this PR. You must list at least one issue.*

*If you had to change anything in the [flutter/tests] repo, include a link to the migration guide as per the [breaking change policy].*