Follow redirects for ktlint download (#3543)
The github release page is redirecting, so follow the redirects. 

Good news, the uploading is working and the cipd package is created: https://chrome-infra-packages.appspot.com/p/flutter/ktlint/linux-amd64

Bad news, the uploaded ktlint file is empty because the curl is failing because we aren't following the redirect. 

Passing the `-L` to curl fixes this (I double checked that with `-L` passed locally, we download a working version of ktlint). Sorry, should have caught this in the original PR.