Reland buildbucket v2 migration parts 3&4 (#3697)
Reverts https://github.com/flutter/cocoon/pull/3686, re-landing the most recent parts of the buildbucket v2 migration.

The one difference relative to the original PRs (see commit 2 in the PR) is that in addition to duplicating all the services and handlers, the original PR duplicated the *cron jobs*. My initial impression of the scheduler code is that there's a non-trivial risk of race conditions in the current design, and so my speculation is that having two cron jobs scheduled at the same time, both trying to run backfill, tripped race conditions that put the databases into bad states.

This is definitely speculative, but given that my understanding is that v2 was successfully handling traffic for at least a day, a race condition rather than a fundamental problem with the implementation of the v2 communication protocol is certainly plausible. My proposal is that we deploy this to test that theory, and to get new data for failure analysis if it doesn't work, but be prepared to roll back and manually adjust the DB again if necessary.