Vacuum stale In Progress tasks more frequently
https://github.com/flutter/flutter/issues/122117 is still happening
frequently. Currently the cron job is only running every 12 hours, which
means that when it happens we can have large numbers of stuck columns on
the dashboard (it seems that when this happens it sometimes happens to
an entire batch of backfill scheduling, which is currently 50) for up to
12 hours.

Looking at the history of this task:
- Originally it ran every 10 minutes, and was much more aggressive in what
  tasks it reset, causing https://github.com/flutter/flutter/issues/121989
  and was subsequently disabled in https://github.com/flutter/cocoon/pull/2515
- It was re-enabled in https://github.com/flutter/cocoon/pull/3354, more
  narrowly targetted so it wouldn't reset jobs that had been assigned
  builds. However, it was re-landed with a much more conservative 12
  hour cycle. That version *replaced* the time-based check, however,
  instead of adding to it, which means it's subject to a race when it
  runs. The justification for removing it was that since the job
  frequency was low the race wouldn't be an issue (which is only
  somewhat true).

This re-adds a time-based check to avoid races, and increases the
frequency to one hour (which may still cause some tree delays; we can
evaluate once it's been running in this mode for a while if we want to
make it more often and make the timestamp check less conservative).
It also removes the reset of the creation timestamp to 0 so that the
timestamp check will work; that appears to be a legacy of the earliest
version of this job which was trying to clean up a slightly different
case involving null creation timestamps.

Mitigation for https://github.com/flutter/flutter/issues/122117