Remove most Buildbucket v1 code (#3700)
This eliminates most of the Buildbucket v1 code:
- Removes a number of v1 REST handlers with the v2 version (making the two versioned endpoints for these actually call identical code internally):
  - batch backfiller serviceâthe code was identical other than using the v2 scheduler
  - github vacuum serviceâthe code was identical other than using the v2 scheduler
  - try job reset serviceâthe code was identical other than using the v2 scheduler
  - prod job reset serviceâthe code was identical other than using the v2 scheduler, and the v2 version didn't get the logging from https://github.com/flutter/cocoon/pull/3636 so I copied that over.
- Removes the v1 scheduler abstraction, replacing all use with the v2 version.
- Removes the unused portions of the v1 GitHub checks service.
- Removes the v1 LCUI build service abstraction, replacing almost all use with the v2 version.
- Removes the no-longer-used v1 Buildbucker methods (in buildbucket.dart)

The v1 code that remains is code for which v2 services can't be used as a drop-in replacement, which is places where the legacy v1 pubsub subscription handlers take incoming v1 messages and use them to reschedule builds. To replace that with v2 would require writing a complex v1-data-to-v2-data translation layer, which shouldn't be necessary because nothing should be coming in on the v1 pubsub subscription in the first place. However, because that's harder to reason directly about (since it depends on more than the local code), I'm leaving removing that for a later follow-up step; once this PR is deployed, we can guarantee that we aren't accidentally generating any new v1 requests, so then we can wait, watch server traffic, and then do a follow-up to simply remove the v1 subscription logic and all the code it uses when we see that there is in fact none. (Since that will be a separate PR, it will also be easy to roll back via traffic redirection if we somehow miss something.)

Even where the v1 versions of classes are removed, the v2 class and file naming is left for now. Cleaning that will be done in a later follow-up since diffs will then correctly show it as a move, whereas if I moved them over the deleted code in the same PR, it would look like this PR was deleting the v2 version and doing major rewrites to v1, making the PR much harder to reason about.

Part of https://github.com/flutter/flutter/issues/135934