Replace installation query with user-specific version
Instead of querying all installations and looking for the one owned by a
specific user, use the endpoint that allows requesting the installation
for a specific user. This avoids issues with pagination, as well as
making the logic for processing the result simpler.

Fixes https://github.com/flutter/flutter/issues/148092