Fix presubmit retry of missing builder (#3708)
Currently the presubmit LUCI pubsub handler has a couple of un-ACK'd messages because the PR has a ci.yaml that references a builder that was recently removed. The current code to determine whether to retry the job is throwing an uncaught exception when validating the config, causing the handler to return a 500.

This catches that error, and forces the job not to retry, so that the message can be processed. Not retrying is the correct behavior, since it would fail again anway.