Scheduler test bringup (#3779)
Uncomment and finish porting several scheduler tests.

While working on updating https://github.com/flutter/cocoon/pull/3675 to v2, I discovered that a number of `scheduler_test.dart` tests had been commented out in the v2 copy of the file (now the only copy). This uncomments all commented-out tests in this file, and finishes converting them to the v2 APIs. This was a fairly mechanical process, so I expect that this was just an oversight in the v2 bringup.

This required some minor type adjustments in `luci_build_service.dart`, but other code in that same file already uses `<String, Object?>` for properties instead, so this change seems reasonable.