dart_test_runner: output expected dart format args
Currently on a failed run, the output looks like:

    ######### dart format #########
    Formatted lib/src/file_codesign_visitor.dart
    Formatted 10 files (1 changed) in 0.65 seconds.

In the face of this output, my natural inclination would be to run:

    dart format path/to/file

but that would be incorrect. What would be correct would be to locate our CI test script, read the code, and discover that we format to a width of 120 characters, then format with those args.

Instead, we now inform the user of the args directly in the test output to save them some digging.