dart_test_runner: output expected dart format args (#3889)
Currently on a failed run due to unformatted dart code, the output looks like:

    ######### dart format #########
    Formatted lib/src/file_codesign_visitor.dart
    Formatted 10 files (1 changed) in 0.65 seconds.

In the face of this output, my natural inclination would be to run

    dart format path/to/file

but that would be incorrect. What would be correct would be to locate our CI test script, read the code, and discover that we format to a width of 120 characters, then format with those args. 

Instead, we now inform the user of the args directly in the test output to save them some digging.

Also checks to see the user actually supplied an argument to the script and if not, emits an error message and usage info then fails, rather than silently running  `pushd ""`, resulting in no change to PWD, then failing when it realises that the top level of the repo is not a pub package.