Remove unsafe_html lint rule
The unsafe_html lint rule was not meant for external usage. It was implemented
for internal usage, before there was a set of internal lint rules.

This rule is not maintained, and does not have any external design. It is also
meant to be used with cannot-ignore and in a VCS where specific owners have
rights to specific files. None of this applies to this repo. Additionally, the
rule only helps with dart:html elements, which is a library that is not
imported anywhere in this repository.

See also dart-lang/linter#5001 and flutter/flutter#157530