use minimal number of arguments to enqueue PRs (#4012)
This is a speculative fix for `enqueuePullRequest` internal error:

```
OperationException(linkException: null, graphqlErrors: [GraphQLError(message: Could not resolve to a node with the global id of '2157301061', locations: [ErrorLocation(line: 2, column: 3)], path: [enqueuePullRequest], extensions: null)])"
```

It's unclear what "global id of '2157301061'" is and where it comes from (we're not passing it). The hypothesis in this PR is that perhaps Github is failing to lookup the `expectedHeadOid` or something in the return value, and so reducing the argument list to just the essential one and not requesting anything to be returned might fix the issue.