[mq] do not enqueue already enqueued PRs (#4017)
A PR stays open while it's in the merge queue because the MQ can eject it back if it doesn't pass tests, so `PullRequestValidationService` can receive pings that would look like a request to validate and merge the PR (it will be all green, reviewed, and have an `autosubmit` label). However, if a PR is already in the merge queue, there's nothing to do. So short-circuit the validation process, do nothing other than print a log message that it happened.