CiStaging Firestore model + test (#4018)
Define the interface for "staging" documents in Cocoon. These are documents in firestore that will record different stages in a monorepo such as the engine-build stage (required before we can run tests).

Previously ci.yaml validation was a single stage: "can cocoon schedule all the build tasks"
In a monorepo we will break this up into multiple stages, with tasks filling in the fields of the firebase document.

Take note; this must be done with a transaction read/write. We will be told by github/luic about jobs being finished asynchronously. Two tasks completing at the same time would set the "remaining" work to the same value and break.